Kodiak Drake 9/7/2022
In MrRoboto, created a program that checks for a certain input, and with the correct input, it gives an output.

In GettingHotInHere, created a program where one can input a fahrenheit value, and it gives an output of fahrenheit, celsius, as a double and an integer, and also tells the temp in °C if it were 2° higher.